final group/XP code - group corpses - group kil range - group quest item drops - several fixes

+++ server/src/socket/image.c
@@ -300,7 +300,7 @@ void SetFaceMode(char *buf, int len, NewSocket *ns)
 void SendFaceCmd(char *buff, int len, NewSocket *ns)
 {
     long    tmpnum  = atoi(buff);
-    short   facenum = tmpnum & 0xffff;
+    short   facenum = (short)(tmpnum & 0xffff);
 
     if (facenum != 0)
         esrv_send_face(ns, facenum, 1);
